+++ common/src/main/java/mb/common/result/ThrowingPredicate.java
@@ -1,0 +1,6 @@
+package mb.common.result;
+
+@FunctionalInterface
+public interface ThrowingPredicate<T, E extends Exception> {
+    boolean test(T t) throws E;
+}