+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project are documented in this file, based on [Keep 
 ## [Unreleased]
 
 
+## [0.9.6] - 2021-09-24
+### Added
+- `flatMapThrowing`, `filter`, `filterThrowing`, `stream` to `Option`.
+- `ThrowingPredicate`, a throwing version of `Predicate`.
+
+
 ## [0.9.5] - 2021-09-20
 ### Added
 - `tryLock` methods to `CloseableReentrantReadWriteLock` and `LockHandle`
@@ -21,7 +27,8 @@ All notable changes to this project are documented in this file, based on [Keep 
 - Update resource dependency to 0.11.5.
 
 
-[Unreleased]: https://github.com/metaborg/common/compare/release-0.9.5...HEAD
+[Unreleased]: https://github.com/metaborg/common/compare/release-0.9.6...HEAD
+[0.9.6]: https://github.com/metaborg/common/compare/release-0.9.5...release-0.9.6
 [0.9.5]: https://github.com/metaborg/common/compare/release-0.9.4...release-0.9.5
 [0.9.4]: https://github.com/metaborg/common/compare/release-0.9.3...release-0.9.4
 [0.9.3]: https://github.com/metaborg/common/compare/release-0.9.2...release-0.9.3